+++ R/read_lines.R
@@ -15,6 +15,9 @@
 read_lines <- function(file, skip = 0, n_max = -1L,
                        locale = default_locale(),
                        progress = interactive()) {
+  if (empty_file(file)) {
+    return(character())
+  }
   ds <- datasource(file, skip = skip)
   read_lines_(ds, locale_ = locale, n_max = n_max, progress = progress)
 }