+++ NEWS.md
@@ -11,6 +11,8 @@
   compatibility with the `widths` argument in `read.fwf()`. (#380, @leeper)
 * `type_covert()` now accepts only `NULL` or a `cols` specification for
   `col_types` (#369, @jimhester).
+* `read_file()`, `read_lines()` and `read_csv()` now return empty objects
+  rather than signaling an error when run on an empty file (#356, @jimhester).
 
 * time objects returned by `parse_time()` are now `hms` objects rather than a
   custom `time` class (#409, @jimhester).